+++ pkgs/dart_mcp/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update the tool calling example to include progress notifications.
 - Remove a reference to "screenshot" for a generic error that occurs for more
   than just screenshots.
+- Mark the "root" parameter for create_project required.
 
 ## 0.3.3